+++ pipeline.py
@@ -1,12 +1,12 @@
-from options.infer_options import InferOptions
 import numpy as np
 import torch
+from keras.models import load_model
 from PIL import Image
 from torchvision import transforms
-from keras.models import load_model
 
 from models.PINet20 import TransferModel, create_model
-from tool import cords_to_map, get_coords
+from options.infer_options import InferOptions
+from tool import cords_to_map, get_coords, reorder_pose, load_pose_from_file
 from util import util
 
 IMAGE_SIZE = (256, 176)
@@ -23,7 +23,16 @@ class InferencePipeline:
             transforms.Normalize((0.5, 0.5, 0.5), (0.5, 0.5, 0.5))
         ])
 
-    def __call__(self, image: Image, target_pose_map) -> Image:
+    @classmethod
+    def from_opts(cls, opt):
+        TEST_SEG_PATH = 'test_data/testSPL2/randomphoto_small.png'
+
+        pinet = create_model(opt)
+        pose_estimator = load_model(opt.pose_estimator, compile=False)
+        segmentator = DummySegmentationModel(TEST_SEG_PATH)
+        return cls(pose_estimator, pinet, segmentator, opt)
+
+    def __call__(self, image: Image, target_pose_map: torch.Tensor) -> Image:
         # get pose
         imgBGR = np.array(image)[:, :, ::-1]
         pose = get_coords(imgBGR, self.pose_estimator)
@@ -36,7 +45,7 @@ class InferencePipeline:
 
         # run PINet
         image_norm = self.transform(image).unsqueeze(0)
-        
+
         if self.opt.gpu_ids:
             # move data to GPU
             device = self.opt.gpu_ids[0]
@@ -44,9 +53,10 @@ class InferencePipeline:
             target_pose_map = target_pose_map.cuda(device)
             image_norm = image_norm.cuda(device)
             spl_onehot = spl_onehot.cuda(device)
-            
-        output_image, output_segmentation = self.pinet.infer(image_norm, pose_map, target_pose_map, spl_onehot)
-        return output_image, output_segmentation
+
+        output_image, output_segmentation = self.pinet.infer(
+            image_norm, pose_map, target_pose_map, spl_onehot)
+        return Image.fromarray(util.tensor2im(output_image))
 
 
 class DummySegmentationModel:
@@ -68,31 +78,18 @@ class DummySegmentationModel:
         # SPL = torch.from_numpy(SPL_img).long()
         return SPL_onehot
 
-def reorder_pose(pose_map):
-    pose = torch.from_numpy(pose_map).float()  # h, w, c
-    pose = pose.transpose(2, 0)  # c,w,h
-    pose = pose.transpose(2, 1)  # c,h,w
-    return pose.unsqueeze(0)
-
-def load_pose_from_file(path):
-    pose_img = np.load(path)
-    return reorder_pose(pose_img)
 
 if __name__ == '__main__':
     SOURCE_IMAGE_PATH = 'test_data/test/randomphoto_small.jpg'
-    TEST_SEG_PATH = 'test_data/testSPL2/randomphoto_small.png'
     TARGET_POSE_PATH = 'test_data/testK/randomphoto_small.jpg.npy'
     OUPUT_PATH = 'test_data/out.jpg'
 
     opt = InferOptions().parse()
-    pinet = create_model(opt)
-    pose_estimator = load_model(opt.pose_estimator, compile=False)
-    segmentator = DummySegmentationModel(TEST_SEG_PATH)
-    pipeline = InferencePipeline(pose_estimator, pinet, segmentator, opt)
+    pipeline = InferencePipeline.from_opts(opt)
 
     source_image = Image.open(SOURCE_IMAGE_PATH)
     target_pose = load_pose_from_file(TARGET_POSE_PATH)
 
-    output_image, output_segmentation = pipeline(source_image, target_pose)
+    output_image = pipeline(source_image, target_pose)
 
-    Image.fromarray(util.tensor2im(output_image)).save(OUPUT_PATH)
+    output_image.save(OUPUT_PATH)
